tests/wpt/web-platform-tests/workers/interfaces/WorkerGlobalScope/close/incoming-message.html

<!--
onmessage = function(e) {
  postMessage(1);
  throw new Error();
}
close();
/*
-->
<!doctype html>
<title>close() and incoming message</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var gotMessage = false;
  var gotError = false;
  var worker = new Worker('#');
  worker.onmessage = function(e) {
    gotMessage = true;
  };
  worker.onerror = function(e) {
    gotError = true;
  };
  worker.postMessage(1);
  setTimeout(this.step_func(function() {
    assert_false(gotMessage, 'got message');
    assert_true(gotError, 'got error');
    this.done();
  }), 500);
});
</script>
<!--
*/
//-->
